Time: 18 ms (47.83%), Space: 47 MB (81.16%) - LeetHub

+++ 0943-sum-of-subarray-minimums/0943-sum-of-subarray-minimums.cpp
@@ -36,6 +36,6 @@ public:
             sum = (sum + arr[i] * left * right) % MOD;
         }
 
-        return (int)sum;
+        return sum;
     }
 };
